Replace signal.h with only auto_signal.h to avoid system conflicts
